+++ src/teletask-test-card.ts
@@ -132,7 +132,7 @@ export class TeletaskTestCard extends LitElement {
 });
 
 console.info(
-  '%c TELETASK-TEST-CARD %c v1.8.1 ',
+  '%c TELETASK-TEST-CARD %c v1.9.0 ',
   'background-color: #03a9f4; color: #fff; font-weight: bold;',
   'background-color: #333; color: #fff; font-weight: bold;'
 );